Fixup of Back/Cancel buttons for older LPA workflow
This effectively locks in the new workflow as if the feature flag was permanently turned on and will break functionality if the flag is ever reverted. A future piece of work will need doing to ensure that feature flags can be used in templates to provide switching of html elements.