update contact-details template to not have show options. Individual templates made for each contact type
contact-details has been split to an address, email and telephone template.

Contact-details now references all three of these templates so we can include this template when all details need to be present.

For cases where specific ordering is needed or only some contact-details are needed to be shown individual templates should be added