Unfortunately large commit that should probably have not happened
Add authorisation middleware that pulls claims from a passed JWT and makes it available in the request context.

Pull configuration out to the top level main.go and ensure it's passed down as method parameters as it should be.

Utilise AWS SDK v2 and make changes to allow context usage with that.