Refactor user id fingerprinting
Move much of the logic around selection of headers into a domain specific object and the working of that object to the appropriate service, not the handler.

Additionally, in the middleware, check to see if the route being accessed is a javascript one and stop processing of the id at that point.